+++ test/outline/test_outline.py
@@ -3,6 +3,7 @@ from pathlib import Path
 import pytest
 
 from fpdf import FPDF, TextStyle, TitleStyle, errors
+from fpdf.enums import Align
 from fpdf.outline import TableOfContents
 
 from test.conftest import LOREM_IPSUM, assert_pdf_equal
@@ -65,6 +66,31 @@ def test_incoherent_start_section_hierarchy():
     pdf.start_section("Title", level=0)
     with pytest.raises(ValueError):
         pdf.start_section("Subtitle", level=2)
+
+
+def test_start_section_horizontal_alignment(tmp_path):  # issue-1282
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("Helvetica", "", 20)
+
+    # left align
+    level0 = TextStyle("Helvetica", "", 20, (0, 0, 0), l_margin=Align.L)
+    pdf.set_section_title_styles(level0)
+    pdf.start_section("left aligned section")
+
+    # center align
+    level0 = TextStyle("Helvetica", "", 20, (0, 0, 0), l_margin=Align.C)
+    pdf.set_section_title_styles(level0)
+    pdf.start_section("center aligned section")
+
+    # right align
+    level0 = TextStyle("Helvetica", "", 20, (0, 0, 0), l_margin=Align.R)
+    pdf.set_section_title_styles(level0)
+    pdf.start_section("right aligned section")
+
+    assert_pdf_equal(
+        pdf, HERE / "test_start_section_horizontal_alignment.pdf", tmp_path
+    )
 
 
 def test_set_section_title_styles_with_invalid_arg_type():
@@ -522,7 +548,6 @@ def test_toc_extra_pages_with_labels(tmp_path):
         pdf.cell(text=pdf.get_page_label(), center=True)
 
     for test_number in range(3):
-
         pdf = FPDF()
         pdf.footer = footer
 
@@ -584,13 +609,10 @@ def test_toc_extra_pages_with_labels(tmp_path):
             pdf.ln()
             pdf.add_font(
                 family="Quicksand",
-                style="",
                 fname=HERE.parent / "fonts" / "Quicksand-Regular.otf",
             )
             toc = TableOfContents()
-            toc.text_style = TextStyle(
-                font_family="Quicksand", font_style="", font_size_pt=14
-            )
+            toc.text_style = TextStyle(font_family="Quicksand", font_size_pt=14)
             pdf.insert_toc_placeholder(toc.render_toc, allow_extra_pages=True)
 
         if test_number == 2:
@@ -605,8 +627,7 @@ def test_toc_extra_pages_with_labels(tmp_path):
             )
             pdf.ln()
             pdf.ln()
-            toc = TableOfContents()
-            toc.use_section_title_styles = True
+            toc = TableOfContents(use_section_title_styles=True)
             pdf.insert_toc_placeholder(toc.render_toc, allow_extra_pages=True)
 
         pdf.set_page_label(label_style="D")